More precise calculations

+++ src/main/java/com/github/dirtpowered/releasetobeta/network/translator/betatomodern/EntityTeleportTranslator.java
@@ -13,9 +13,9 @@ public class EntityTeleportTranslator implements BetaToModern<EntityTeleportPack
     @Override
     public void translate(EntityTeleportPacketData packet, BetaClientSession session, Session modernSession) {
         int entityId = packet.getEntityId();
-        int x = Utils.toModernPos(packet.getX());
-        int y = Utils.toModernPos(packet.getY());
-        int z = Utils.toModernPos(packet.getZ());
+        double x = Utils.toModernPos(packet.getX());
+        double y = Utils.toModernPos(packet.getY());
+        double z = Utils.toModernPos(packet.getZ());
 
         float yaw = packet.getYaw();
         float pitch = packet.getPitch();
